EZKL upgrade to v18.1.5 (#53)
