--- neurons/deployment_layer/model_1d60d545b7c5123fd60524dcbaf57081ca7dc4a9ec36c892927a3153328d17c0/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "circuit"
-version = "0.1.0"
-edition = "2021"
-
-[workspace]
-members = ["guest"]
-
-[profile.release]
-debug = 1
-codegen-units = 1
-lto = "fat"
-
-[dependencies]
-jolt-sdk = { git = "https://github.com/a16z/jolt", features = ["host"], rev = "9f0b9e6d95814dfe15d74ea736b9f89d505e8d07" }
-jolt-core = { git = "https://github.com/a16z/jolt", features = ["host"], rev = "9f0b9e6d95814dfe15d74ea736b9f89d505e8d07" }
-guest = { path = "./guest" }
-print-no-std = "0.1.0"
-clap = "4.5.16"
-serde = { version = "1.0", features = ["derive"], default-features = false }
-serde_json = "1.0.125"
-postcard = "1.0.10"
-log = "0.4.22"
-env_logger = "0.11.5"
-serde-big-array = "0.5.1"
-
-[dev-dependencies]
-tempfile = "3.8.1"
-
-[patch.crates-io]
-ark-ff = { git = "https://github.com/a16z/arkworks-algebra", branch = "optimize/field-from-u64" }
-ark-ec = { git = "https://github.com/a16z/arkworks-algebra", branch = "optimize/field-from-u64" }
-ark-serialize = { git = "https://github.com/a16z/arkworks-algebra", branch = "optimize/field-from-u64" }
