--- neurons/deployment_layer/model_1d60d545b7c5123fd60524dcbaf57081ca7dc4a9ec36c892927a3153328d17c0/guest/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "guest"
-version = "0.1.0"
-edition = "2021"
-
-[[bin]]
-name = "guest"
-path = "./src/lib.rs"
-
-[features]
-guest = []
-
-[dependencies]
-jolt = { package = "jolt-sdk", git = "https://github.com/a16z/jolt", rev = "9f0b9e6d95814dfe15d74ea736b9f89d505e8d07" }
-serde = { version = "1.0.209", default-features = false }
-serde-big-array = "0.5.1"
-libm = "0.2.8"
