--- neurons/deployment_layer/model_1d60d545b7c5123fd60524dcbaf57081ca7dc4a9ec36c892927a3153328d17c0/metadata.json
@@ -1,11 +1,0 @@
-{
-    "name": "SN27 PoW 256",
-    "description": "Generates weights based on miner evaluation data from SN27.",
-    "author": "Inference Labs",
-    "version": "0.0.5",
-    "proof_system": "JOLT",
-    "netuid": 27,
-    "type": "proof_of_weights",
-    "external_files": {},
-    "benchmark_choice_weight": 0
-}
