--- neurons/deployment_layer/model_33b92394b18412622adad75733a6fc659b4e202b01ee8a5465958a6bad8ded62/metadata.json
@@ -1,14 +1,0 @@
-{
-    "name": "SN48 PoW",
-    "description": "Generates weights based on miner scoring data from SN48.",
-    "author": "Inference Labs",
-    "version": "0.0.1",
-    "proof_system": "EZKL",
-    "netuid": 48,
-    "weights_version": 0,
-    "type": "proof_of_weights",
-    "external_files": {
-        "pk.key": "https://storage.omron.ai/33b92394b18412622adad75733a6fc659b4e202b01ee8a5465958a6bad8ded62/pk.key"
-    },
-    "benchmark_choice_weight": 0.2
-}
