--- neurons/deployment_layer/model_37320fc74fec80805eedc8e92baf3c58842a2cb2a4ae127ad6e930f0c8441c7a/metadata.json
@@ -1,10 +1,0 @@
-{
-    "name": "SN2 PoW 256",
-    "description": "Generates weights based on miner evaluation data from SN2.",
-    "author": "Inference Labs",
-    "version": "0.0.5",
-    "proof_system": "JOLT",
-    "type": "proof_of_computation",
-    "external_files": {},
-    "benchmark_choice_weight": 0
-}
