+++ neurons/deployment_layer/model_50818a54b31b3e0fe3306a7fb7826156fc2c42c9d64c6ba106ba135fbe7b7b19/metadata.json
@@ -1,0 +1,12 @@
+{
+    "name": "LSTM",
+    "description": "Default LSTM benchmarker model.",
+    "author": "Inference Labs",
+    "version": "0.0.2",
+    "proof_system": "EZKL",
+    "type": "proof_of_computation",
+    "external_files": {
+        "pk.key": "https://storage.omron.ai/50818a54b31b3e0fe3306a7fb7826156fc2c42c9d64c6ba106ba135fbe7b7b19/pk.key"
+    },
+    "benchmark_choice_weight": 0.2
+}
