+++ neurons/deployment_layer/model_50818a54b31b3e0fe3306a7fb7826156fc2c42c9d64c6ba106ba135fbe7b7b19/settings.json
@@ -1,0 +1,95 @@
+{
+    "run_args": {
+        "tolerance": {
+            "val": 0.0,
+            "scale": 1.0
+        },
+        "input_scale": 13,
+        "param_scale": 13,
+        "scale_rebase_multiplier": 1,
+        "lookup_range": [
+            -8582,
+            21116
+        ],
+        "logrows": 18,
+        "num_inner_cols": 2,
+        "variables": [
+            [
+                "batch_size",
+                1
+            ]
+        ],
+        "input_visibility": "Public",
+        "output_visibility": "Public",
+        "param_visibility": "Fixed",
+        "rebase_frac_zero_constants": false,
+        "check_mode": "UNSAFE",
+        "commitment": "KZG",
+        "decomp_base": 16384,
+        "decomp_legs": 2,
+        "bounded_log_lookup": false
+    },
+    "num_rows": 244209,
+    "total_assignments": 488418,
+    "total_const_size": 3259,
+    "total_dynamic_col_size": 0,
+    "max_dynamic_input_len": 0,
+    "num_dynamic_lookups": 0,
+    "num_shuffles": 0,
+    "total_shuffle_col_size": 0,
+    "model_instance_shapes": [
+        [
+            1,
+            5,
+            1
+        ],
+        [
+            1,
+            1
+        ]
+    ],
+    "model_output_scales": [
+        13
+    ],
+    "model_input_scales": [
+        13
+    ],
+    "model_input_types": [
+        "F32"
+    ],
+    "module_sizes": {
+        "polycommit": [],
+        "poseidon": [
+            0,
+            [
+                0
+            ]
+        ]
+    },
+    "required_lookups": [
+        {
+            "Sigmoid": {
+                "scale": 8192.0
+            }
+        },
+        {
+            "Tanh": {
+                "scale": 8192.0
+            }
+        }
+    ],
+    "required_range_checks": [
+        [
+            -1,
+            1
+        ],
+        [
+            0,
+            16383
+        ]
+    ],
+    "check_mode": "UNSAFE",
+    "version": "18.1.5",
+    "num_blinding_factors": null,
+    "timestamp": 1738372231183
+}
