--- neurons/deployment_layer/model_8dcff627a782525ea86196941a694ffbead179905f0cd4550ddc3df9e2b90924/metadata.json
@@ -1,12 +1,0 @@
-{
-    "name": "LSTM",
-    "description": "Default LSTM benchmarker model.",
-    "author": "Inference Labs",
-    "version": "0.0.1",
-    "proof_system": "EZKL",
-    "type": "proof_of_computation",
-    "external_files": {
-        "pk.key": "https://storage.omron.ai/8dcff627a782525ea86196941a694ffbead179905f0cd4550ddc3df9e2b90924/pk.key"
-    },
-    "benchmark_choice_weight": 0.2
-}
