--- neurons/deployment_layer/model_8dcff627a782525ea86196941a694ffbead179905f0cd4550ddc3df9e2b90924/settings.json
@@ -1,95 +1,0 @@
-{
-    "run_args": {
-        "tolerance": {
-            "val": 0.0,
-            "scale": 1.0
-        },
-        "input_scale": 13,
-        "param_scale": 13,
-        "scale_rebase_multiplier": 1,
-        "lookup_range": [
-            -4394,
-            19162
-        ],
-        "logrows": 18,
-        "num_inner_cols": 2,
-        "variables": [
-            [
-                "batch_size",
-                1
-            ]
-        ],
-        "input_visibility": "Public",
-        "output_visibility": "Public",
-        "param_visibility": "Fixed",
-        "rebase_frac_zero_constants": false,
-        "check_mode": "UNSAFE",
-        "commitment": "KZG",
-        "decomp_base": 16384,
-        "decomp_legs": 2,
-        "bounded_log_lookup": false
-    },
-    "num_rows": 157145,
-    "total_assignments": 314290,
-    "total_const_size": 3259,
-    "total_dynamic_col_size": 0,
-    "max_dynamic_input_len": 0,
-    "num_dynamic_lookups": 0,
-    "num_shuffles": 0,
-    "total_shuffle_col_size": 0,
-    "model_instance_shapes": [
-        [
-            1,
-            5,
-            1
-        ],
-        [
-            1,
-            1
-        ]
-    ],
-    "model_output_scales": [
-        13
-    ],
-    "model_input_scales": [
-        13
-    ],
-    "model_input_types": [
-        "F32"
-    ],
-    "module_sizes": {
-        "polycommit": [],
-        "poseidon": [
-            0,
-            [
-                0
-            ]
-        ]
-    },
-    "required_lookups": [
-        {
-            "Tanh": {
-                "scale": 8192.0
-            }
-        },
-        {
-            "Sigmoid": {
-                "scale": 8192.0
-            }
-        }
-    ],
-    "required_range_checks": [
-        [
-            0,
-            16383
-        ],
-        [
-            -1,
-            1
-        ]
-    ],
-    "check_mode": "UNSAFE",
-    "version": "16.2.0",
-    "num_blinding_factors": null,
-    "timestamp": 1733868931658
-}
