--- neurons/deployment_layer/model_a4bcecaf699fd9212600a1f2fcaa40c444e1aeaab409ea240a38c33ed356f4e2/metadata.json
@@ -1,14 +1,0 @@
-{
-    "name": "SN27 PoW 256",
-    "description": "The validator's reward model for SN27, designed to batch a single round of weight updates (256 changes per proof).",
-    "author": "Inference Labs",
-    "version": "0.0.6",
-    "proof_system": "EZKL",
-    "netuid": 27,
-    "weights_version": 160,
-    "type": "proof_of_weights",
-    "external_files": {
-        "pk.key": "https://storage.omron.ai/a4bcecaf699fd9212600a1f2fcaa40c444e1aeaab409ea240a38c33ed356f4e2/pk.key"
-    },
-    "benchmark_choice_weight": 0.2
-}
