+++ neurons/deployment_layer/model_ff9f4069387597719c74a936252aa7e441194d0bb8d422a59c3a581245ad7989/settings.json
@@ -1,0 +1,119 @@
+{
+    "run_args": {
+        "tolerance": {
+            "val": 0.0,
+            "scale": 1.0
+        },
+        "input_scale": 20,
+        "param_scale": 20,
+        "scale_rebase_multiplier": 1,
+        "lookup_range": [
+            0,
+            0
+        ],
+        "logrows": 21,
+        "num_inner_cols": 2,
+        "variables": [
+            [
+                "batch_size",
+                1
+            ]
+        ],
+        "input_visibility": "Public",
+        "output_visibility": "Public",
+        "param_visibility": "Fixed",
+        "rebase_frac_zero_constants": false,
+        "check_mode": "UNSAFE",
+        "commitment": "KZG",
+        "decomp_base": 16384,
+        "decomp_legs": 2,
+        "bounded_log_lookup": false
+    },
+    "num_rows": 901510,
+    "total_assignments": 1803020,
+    "total_const_size": 8,
+    "total_dynamic_col_size": 0,
+    "max_dynamic_input_len": 0,
+    "num_dynamic_lookups": 0,
+    "num_shuffles": 0,
+    "total_shuffle_col_size": 0,
+    "model_instance_shapes": [
+        [
+            256
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            256
+        ],
+        [
+            1
+        ]
+    ],
+    "model_output_scales": [
+        20,
+        0
+    ],
+    "model_input_scales": [
+        20,
+        20,
+        20,
+        20,
+        20,
+        20,
+        0
+    ],
+    "model_input_types": [
+        "F32",
+        "F32",
+        "F32",
+        "F32",
+        "F32",
+        "F32",
+        "Int"
+    ],
+    "module_sizes": {
+        "polycommit": [],
+        "poseidon": [
+            0,
+            [
+                0
+            ]
+        ]
+    },
+    "required_lookups": [],
+    "required_range_checks": [
+        [
+            0,
+            1048575
+        ],
+        [
+            -1,
+            1
+        ],
+        [
+            0,
+            16383
+        ]
+    ],
+    "check_mode": "UNSAFE",
+    "version": "18.1.5",
+    "num_blinding_factors": null,
+    "timestamp": 1738371564833
+}
