+++ pyproject.toml
@@ -9,22 +9,22 @@ description = ""
 readme = "README.md"
 requires-python = "==3.12.*"
 dependencies = [
-	"attrs==24.3.0",
-	"bittensor==8.5.2",
-	"ezkl==16.2.3",
-	"fastapi==0.110.3",
-	"jsonrpcserver>=5.0.9",
-	"numpy==2.0.2",
-	"packaging==24.2",
-	"prometheus_client==0.21.1",
-	"psutil==6.1.1",
-	"pydantic==2.10.6",
-	"pyopenssl>=25.0.0",
-	"requests==2.32.3",
-	"rich==13.8.1",
-	"torch==2.4.1",
-	"uvicorn==0.34.0",
-	"wandb==0.18.7",
+  "attrs==24.3.0",
+  "bittensor==8.5.2",
+  "ezkl==18.1.5",
+  "fastapi==0.110.3",
+  "jsonrpcserver>=5.0.9",
+  "numpy==2.0.2",
+  "packaging==24.2",
+  "prometheus_client==0.21.1",
+  "psutil==6.1.1",
+  "pydantic==2.10.6",
+  "pyopenssl>=25.0.0",
+  "requests==2.32.3",
+  "rich==13.8.1",
+  "torch==2.4.1",
+  "uvicorn==0.34.0",
+  "wandb==0.18.7",
 ]
 
 [tool.setuptools]
@@ -32,7 +32,7 @@ packages = ["omron"]
 package-dir = { "omron" = "neurons" }
 
 [tool.setuptools.dynamic]
-version = {attr = "omron.__version__"}
+version = { attr = "omron.__version__" }
 
 [tool.uv.sources]
 torch = { index = "pytorch" }
@@ -43,6 +43,4 @@ url = "https://download.pytorch.org/whl/cpu"
 explicit = true
 
 [dependency-groups]
-dev = [
-    "black>=24.10.0",
-]
+dev = ["black>=24.10.0"]
